# backend/openapi/escopenote-brain-v1.yaml
openapi: 3.1.0
info:
  title: Escopenote Brain API
  version: 0.1.0
  description: AI gateway for Escopenote desktop (chat SSE, planner).

servers:
  - url: http://127.0.0.1:3001
    description: Local brain dev

paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /v1/chat/stream:
    post:
      summary: Stream chat completion (SSE)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatStreamRequest'
      responses:
        '200':
          description: Server-Sent Events stream
          content:
            text/event-stream:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/ValidationError'

  /v1/planner/generate:
    post:
      summary: Generate study plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlannerGenerateRequest'
      responses:
        '200':
          description: Plan JSON
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlannerGenerateResponse'

components:
  responses:
    ValidationError:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'

  schemas:
    HealthResponse:
      type: object
      required: [ok, version, api, capabilities, provider]
      properties:
        ok: { type: boolean }
        version: { type: string }
        api: { type: string, enum: [v1] }
        capabilities:
          type: array
          items: { type: string }
        provider: { type: string, enum: [gemini, mock] }

    ChatStreamRequest:
      type: object
      required: [message]
      properties:
        message: { type: string, maxLength: 32000 }
        relevant_chunks:
          type: array
          maxItems: 24
          items:
            $ref: '#/components/schemas/RelevantChunk'
        session_id: { type: string }
        course_id: { type: string, nullable: true }
        locale: { type: string }
        feature:
          type: string
          enum: [chat, note_agent, planner, workflow]

    RelevantChunk:
      type: object
      required: [chunkId, fileName, text]
      properties:
        chunkId: { type: string }
        fileName: { type: string }
        text: { type: string, maxLength: 8000 }

    PlannerGenerateRequest:
      type: object
      required: [profile, scope, anchorDate]
      properties:
        profile:
          type: object
          required: [subjects, hoursPerDay, hoursPerWeek]
        scope: { type: string, enum: [day, week] }
        anchorDate: { type: string, format: date }
        locale: { type: string }

    PlannerGenerateResponse:
      type: object
      required: [plan]
      properties:
        plan: { type: object }

    ApiError:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            retryable: { type: boolean }